delete unused useSudokuState. I should have them separate to avoid accidentally reading both when unnecessary
